refactor: reorganize directory structure to align with Next.js conventions
Move non-route files from app/src/app/ to app/src/ to follow Next.js App Router best practices:
- agents/ → src/agents/
- components/ → src/components/
- consts/ → src/consts/
- context/ → src/context/
- libs/ → src/libs/
- schema/ → src/schema/

Update all import paths from @/app/* to @/* accordingly.

The app/src/app/ directory now only contains routes (api/, pages) and root-level files (layout.tsx, page.tsx, globals.css).
