--- app/src/app/consts/index.ts
@@ -1,11 +1,0 @@
-import { ENDPOINT } from "@/app/consts/endpoint";
-import { MESSAGE } from "@/app/consts/message";
-import { PROMPT } from "@/app/consts/prompt";
-import { SETTING } from "@/app/consts/setting";
-
-export const CONSTS = {
-  ENDPOINT,
-  MESSAGE,
-  SETTING,
-  PROMPT
-};
