+++ app/src/components/auth/AuthGuard.tsx
@@ -2,7 +2,7 @@
 import { useRouter } from "next/navigation";
 import { useEffect } from "react";
 
-import { useAuth } from "@/app/context/AuthContext";
+import { useAuth } from "@/context/AuthContext";
 
 /**
  * 認証ガードコンポーネント
