+++ app/src/consts/index.ts
@@ -1,0 +1,11 @@
+import { ENDPOINT } from "@/consts/endpoint";
+import { MESSAGE } from "@/consts/message";
+import { PROMPT } from "@/consts/prompt";
+import { SETTING } from "@/consts/setting";
+
+export const CONSTS = {
+  ENDPOINT,
+  MESSAGE,
+  SETTING,
+  PROMPT
+};
