+++ app/src/consts/prompt.ts
