+++ app/src/consts/setting.ts
