+++ app/src/schema/hearingJsonSchema.ts
